+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.10.0] - [2020-04-09]
+### Added
+- Add Trading Partner API querying capability
+- Request::TradingPartner
+- Response::TradingPartnerData
+- Response::TradingPartnersData
+- Models::TradingPartner
+
 ## [0.9.0] - [2020-04-08]
 ### Added
 - Added EligibilityBenefit#medicare?
@@ -62,10 +70,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Provider
 - Subscriber
 - Encounter
-- Eligibility 
+- Eligibility
 - Authentication
 - Configuration
 
+[0.10.0]: https://github.com/WeInfuse/change_health/compare/v0.9.0...v0.10.0
+[0.9.0]: https://github.com/WeInfuse/change_health/compare/v0.8.0...v0.9.0
 [0.8.0]: https://github.com/WeInfuse/change_health/compare/v0.7.0...v0.8.0
 [0.7.0]: https://github.com/WeInfuse/change_health/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/WeInfuse/change_health/compare/v0.5.0...v0.6.0